chore(build): prevent circular dep, fixes aurelia/new#60
